ui/components.py 中一些专用函数转移至 ui/oneday.py 中
